Implemented Authorization

+++ src/main/java/com/urosdragojevic/realbookstore/controller/PersonsController.java
@@ -1,5 +1,6 @@
 package com.urosdragojevic.realbookstore.controller;
 
+import ch.qos.logback.core.joran.spi.ActionException;
 import com.urosdragojevic.realbookstore.audit.AuditLogger;
 import com.urosdragojevic.realbookstore.domain.Person;
 import com.urosdragojevic.realbookstore.domain.User;
@@ -9,7 +10,11 @@ import jakarta.servlet.http.HttpSession;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -34,6 +39,7 @@ public class PersonsController {
     }
 
     @GetMapping("/persons/{id}")
+    @PreAuthorize("hasAuthority('VIEW_PERSON')")
     public String person(@PathVariable int id, Model model, HttpSession session) {
         model.addAttribute("person", personRepository.get("" + id));
         model.addAttribute("CSRF_TOKEN", session.getAttribute("CSRF_TOKEN"));
@@ -41,6 +47,7 @@ public class PersonsController {
     }
 
     @GetMapping("/myprofile")
+    @PreAuthorize("hasAuthority('VIEW_MY_PROFILE')")
     public String self(Model model, Authentication authentication, HttpSession session) {
         User user = (User) authentication.getPrincipal();
         model.addAttribute("person", personRepository.get("" + user.getId()));
@@ -49,24 +56,43 @@ public class PersonsController {
     }
 
     @DeleteMapping("/persons/{id}")
-    public ResponseEntity<Void> person(@PathVariable int id) {
-        personRepository.delete(id);
-        userRepository.delete(id);
-
+    @PreAuthorize("hasAuthority('UPDATE_PERSON')")
+    public ResponseEntity<Void> person(@PathVariable int id) throws ActionException {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        User usr = (User) auth.getPrincipal();
+        int usrId = usr.getId();
+        boolean hasauth = usr.getAuthorities().contains(new SimpleGrantedAuthority("UPDATE_PERSON"));
+        if(usrId == id || hasauth) {
+            personRepository.delete(id);
+            userRepository.delete(id);
+        }else{
+            throw new ActionException("Forbidden action!");
+        }
         return ResponseEntity.noContent().build();
     }
 
     @PostMapping("/update-person")
-    public String updatePerson(Person person, @RequestParam("csrfToken") String csrfToken, HttpSession session) throws AccessDeniedException {
+    @PreAuthorize("hasAuthority('UPDATE_PERSON')")
+    public String updatePerson(Person person, @RequestParam("csrfToken") String csrfToken, HttpSession session) throws AccessDeniedException, ActionException {
         String csrf = session.getAttribute("CSRF_TOKEN").toString();
         if(!csrf.equals(csrfToken)){
             throw new AccessDeniedException("Invalid CSRF token");
         }
-        personRepository.update(person);
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        User usr = (User) auth.getPrincipal();
+        int usrId = usr.getId();
+        int id = Integer.parseInt(person.getId());
+        boolean hasauth = usr.getAuthorities().contains(new SimpleGrantedAuthority("UPDATE_PERSON"));
+        if(usrId == id || hasauth) {
+            personRepository.update(person);
+        }else{
+            throw new ActionException("Forbidden action!");
+        }
         return "redirect:/persons/" + person.getId();
     }
 
     @GetMapping("/persons")
+    @PreAuthorize("hasAuthority('VIEW_PERSONS_LIST')")
     public String persons(Model model, HttpSession session) {
         model.addAttribute("persons", personRepository.getAll());
         return "persons";
